topic/message interface unified
