+++ Mqtt/test/com/github/pfichtner/ardulink/MqttTest.java
@@ -28,67 +28,6 @@ public class MqttTest {
 
 	private static final String TOPIC = "home/devices/ardulink/";
 
-	public static class Message {
-
-		private final String topic;
-		private final String message;
-
-		public Message(String topic, Object message) {
-			this(topic, String.valueOf(message));
-		}
-
-		public Message(String topic, String message) {
-			this.topic = topic;
-			this.message = message;
-		}
-
-		public String getTopic() {
-			return topic;
-		}
-
-		public String getMessage() {
-			return message;
-		}
-
-		@Override
-		public int hashCode() {
-			final int prime = 31;
-			int result = 1;
-			result = prime * result
-					+ ((message == null) ? 0 : message.hashCode());
-			result = prime * result + ((topic == null) ? 0 : topic.hashCode());
-			return result;
-		}
-
-		@Override
-		public boolean equals(Object obj) {
-			if (this == obj)
-				return true;
-			if (obj == null)
-				return false;
-			if (getClass() != obj.getClass())
-				return false;
-			Message other = (Message) obj;
-			if (message == null) {
-				if (other.message != null)
-					return false;
-			} else if (!message.equals(other.message))
-				return false;
-			if (topic == null) {
-				if (other.topic != null)
-					return false;
-			} else if (!topic.equals(other.topic))
-				return false;
-			return true;
-		}
-
-		@Override
-		public String toString() {
-			return "Message [topic=" + topic + ", message=" + message + "]";
-		}
-
-	}
-
 	private static final String LINKNAME = "testlink";
 
 	private final List<Message> published = new ArrayList<Message>();
@@ -103,14 +42,14 @@ public class MqttTest {
 	private final MqttClient mqttClient = new MqttClient(link,
 			new LinkMessageCallback() {
 				@Override
-				public void publish(String topic, String message) {
+				public void publish(String topic, MqttMessage message) {
 					published.add(new Message(topic, message));
 				}
 
 			});
 
 	{
-		// this is an extremely high coupling of ConnectionContactImpl and Link
+		// there is an extremely high coupling of ConnectionContactImpl and Link
 		// which can not be solved other than injecting the variables through
 		// reflection
 		set(connectionContact, getField(connectionContact, "link"), link);
@@ -178,7 +117,7 @@ public class MqttTest {
 	}
 
 	@Test
-	public void canPowerOnDigitalPin() throws IOException {
+	public void canPowerOnDigitalPin() {
 		mqttClient.messageArrived(TOPIC + "D0/value/set", mqttMessage("true"));
 		assertThat(getMessage(), is("alp://ppsw/0/1\n"));
 	}
@@ -220,7 +159,7 @@ public class MqttTest {
 		mqttClient.publishDigitalPinOnStateChanges(pin);
 		simulateArduinoWrite("alp://dred/" + pin + "/" + value);
 		assertThat(published, is(singletonList(new Message(TOPIC + "D" + pin
-				+ "/value/get", value))));
+				+ "/value/get", mqttMessage(value)))));
 	}
 
 	@Test
@@ -239,7 +178,7 @@ public class MqttTest {
 		mqttClient.publishAnalogPinOnStateChanges(pin);
 		simulateArduinoWrite("alp://ared/" + pin + "/" + value);
 		assertThat(published, is(singletonList(new Message(TOPIC + "A" + pin
-				+ "/value/get", value))));
+				+ "/value/get", mqttMessage(value)))));
 	}
 
 	@Test
@@ -272,8 +211,8 @@ public class MqttTest {
 		return codepoints;
 	}
 
-	private MqttMessage mqttMessage(String message) {
-		return new MqttMessage(message.getBytes());
+	private MqttMessage mqttMessage(Object message) {
+		return new MqttMessage(String.valueOf(message).getBytes());
 	}
 
 	private String getMessage() {
